+++ src/bluez/mod.rs
@@ -1,0 +1,4 @@
+mod client;
+mod proxies;
+
+pub use client::{Bluez as Client, BluezDev as Device};
