+++ src/lib.rs
@@ -4,216 +4,6 @@ use std::fmt::Debug;
 use std::{error, fmt, io};
 use tabled::builder::Builder;
 use tabled::settings::Style;
-use zbus::blocking::Connection;
-use zbus::blocking::fdo::ObjectManagerProxy;
-use zbus::proxy;
-use zbus::zvariant::OwnedObjectPath;
-
-#[proxy(
-    default_service = "org.bluez",
-    default_path = "/org/bluez/hci0",
-    interface = "org.bluez.Adapter1",
-    gen_blocking = true,
-    blocking_name = "BluezAdapterProxy",
-    async_name = "BluezAsyncAdapterProxy"
-)]
-pub trait BluezAdapter {
-    #[zbus(property, name = "PowerState")]
-    fn power_state(&self) -> zbus::Result<String>;
-
-    #[zbus(property)]
-    fn set_powered(&self, power_state: bool) -> zbus::Result<()>;
-}
-
-#[proxy(
-    default_service = "org.bluez",
-    interface = "org.bluez.Device1",
-    gen_blocking = true,
-    blocking_name = "BluezDeviceProxy",
-    async_name = "BluezAsyncDeviceProxy"
-)]
-pub trait BluezDevice {
-    #[zbus(property)]
-    fn connected(&self) -> zbus::Result<bool>;
-
-    #[zbus(property)]
-    fn bonded(&self) -> zbus::Result<bool>;
-
-    #[zbus(property)]
-    fn paired(&self) -> zbus::Result<bool>;
-
-    #[zbus(property)]
-    fn trusted(&self) -> zbus::Result<bool>;
-
-    #[zbus(property)]
-    fn alias(&self) -> zbus::Result<String>;
-
-    #[zbus(property)]
-    fn address(&self) -> zbus::Result<String>;
-}
-
-#[proxy(
-    default_service = "org.bluez",
-    interface = "org.bluez.Battery1",
-    gen_blocking = true,
-    blocking_name = "BluezDeviceBatteryProxy",
-    async_name = "BluezAsyncDeviceBatteryProxy"
-)]
-pub trait BluezDeviceBattery {
-    #[zbus(property)]
-    fn percentage(&self) -> zbus::Result<u8>;
-}
-
-pub enum BluezPowerState {
-    On,
-    Off,
-}
-impl fmt::Display for BluezPowerState {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match &self {
-            BluezPowerState::On => write!(f, "enabled"),
-            BluezPowerState::Off => write!(f, "disabled"),
-        }
-    }
-}
-
-impl From<String> for BluezPowerState {
-    fn from(value: String) -> Self {
-        if &value == "on" {
-            BluezPowerState::On
-        } else {
-            BluezPowerState::Off
-        }
-    }
-}
-
-impl std::ops::Not for BluezPowerState {
-    type Output = Self;
-
-    fn not(self) -> Self::Output {
-        match self {
-            BluezPowerState::On => Self::Off,
-            BluezPowerState::Off => Self::On,
-        }
-    }
-}
-
-impl From<&BluezPowerState> for bool {
-    fn from(value: &BluezPowerState) -> Self {
-        match value {
-            BluezPowerState::On => true,
-            BluezPowerState::Off => false,
-        }
-    }
-}
-
-#[derive(Debug)]
-pub struct BluezDev {
-    alias: String,
-    address: String,
-    connected: bool,
-    paired: bool,
-    trusted: bool,
-    bonded: bool,
-    battery: Option<u8>,
-}
-
-pub struct Bluez {
-    connection: Connection,
-}
-
-impl Bluez {
-    pub fn new() -> zbus::Result<Self> {
-        let connection = Connection::system()?;
-        Ok(Self { connection })
-    }
-
-    fn get_dev_object_paths(&self) -> zbus::Result<Vec<OwnedObjectPath>> {
-        let object_manager_proxy = ObjectManagerProxy::new(&self.connection, "org.bluez", "/")?;
-        let objects = object_manager_proxy.get_managed_objects()?;
-
-        let dev_paths = objects
-            .into_keys()
-            .filter(|k| {
-                if let Some(path) = k.rsplitn(2, "/").take(1).next() {
-                    path.contains("dev")
-                } else {
-                    false
-                }
-            })
-            .collect::<Vec<OwnedObjectPath>>();
-
-        Ok(dev_paths)
-    }
-
-    fn build_proxy<'a, T>(&self, path: Option<&'a str>) -> zbus::Result<T>
-    where
-        T: zbus::blocking::proxy::ProxyImpl<'a> + From<zbus::Proxy<'a>>,
-    {
-        let mut proxy_builder = T::builder(&self.connection);
-
-        if let Some(path) = path {
-            proxy_builder = proxy_builder.path(path)?;
-        }
-
-        proxy_builder.build()
-    }
-
-    pub fn power_state(&self) -> zbus::Result<BluezPowerState> {
-        let adapter_proxy: BluezAdapterProxy = self.build_proxy(None)?;
-        let result = adapter_proxy.power_state().map(BluezPowerState::from)?;
-
-        Ok(result)
-    }
-
-    pub fn toggle_power_state(&self) -> zbus::Result<BluezPowerState> {
-        let adapter_proxy: BluezAdapterProxy = self.build_proxy(None)?;
-        let prev_state = adapter_proxy.power_state().map(BluezPowerState::from)?;
-
-        let new_state = !prev_state;
-        adapter_proxy.set_powered(bool::from(&new_state))?;
-
-        Ok(new_state)
-    }
-
-    pub fn devs(&self) -> zbus::Result<Vec<BluezDev>> {
-        let dev_paths = self.get_dev_object_paths()?;
-
-        Ok(dev_paths
-            .into_iter()
-            .filter_map(|dev_path| {
-                let dev_proxy: BluezDeviceProxy = self.build_proxy(Some(&dev_path)).ok()?;
-
-                let mut dev = BluezDev {
-                    alias: dev_proxy.alias().ok()?,
-                    address: dev_proxy.address().ok()?,
-                    connected: dev_proxy.connected().ok()?,
-                    paired: dev_proxy.paired().ok()?,
-                    trusted: dev_proxy.trusted().ok()?,
-                    bonded: dev_proxy.bonded().ok()?,
-                    battery: None,
-                };
-
-                if !dev.connected {
-                    return Some(dev);
-                }
-
-                let battery_proxy: BluezDeviceBatteryProxy =
-                    self.build_proxy(Some(&dev_path)).ok()?;
-                dev.battery = Some(battery_proxy.percentage().ok()?);
-
-                Some(dev)
-            })
-            .collect::<Vec<BluezDev>>())
-    }
-
-    pub fn connected_devs(&self) -> zbus::Result<Vec<BluezDev>> {
-        let devs = self.devs()?;
-
-        Ok(devs.into_iter().filter(|d| d.connected).collect())
-    }
-}
-
 pub fn status(f: &mut impl io::Write) -> Result<(), Box<dyn error::Error>> {
     let bluez = Bluez::new()?;
 
@@ -380,3 +170,4 @@ pub fn create_terse_out(listing: impl Iterator<Item = Vec<String>>) -> String {
         })
         .collect()
 }
+mod bluez;
